fix(docs): gate Ask AI shortcut on panel presence and guard text inputs
